Added fetching of corp sheet if corporation doesn't exist in database when creating a new user Switched back to mainline redistore repo after selecting a database has been fixed (Temporarily) switched from main eveapi repo to fork in order to use latest additions before pull request Adapted all eveapi calls to use new version

+++ models/account.go
@@ -15,7 +15,7 @@ type Account struct {
 	// APIvCode represents the EVE Online API verification code
 	APIvCode string `json:"apivCode"`
 	// APIAccesMask represents the access mask the EVE Online API key is set to
-	APIAccessMask int `json:"apiAccessMask"`
+	APIAccessMask int64 `json:"apiAccessMask"`
 	// DefaultAccount indicates whether the account is selected as default for the user
 	DefaultAccount bool `json:"defaultAccount"`
 	// Active indicates whether the Account is set as active
@@ -25,7 +25,7 @@ type Account struct {
 }
 
 // NewAccount creates a new account with the given information
-func NewAccount(userID int64, apiKeyID int64, apivCode string, apiAccessMask int, defaultAccount bool, active bool) *Account {
+func NewAccount(userID int64, apiKeyID int64, apivCode string, apiAccessMask int64, defaultAccount bool, active bool) *Account {
 	account := &Account{
 		ID:             -1,
 		UserID:         userID,
